feat: Add account types get endpoint and seeder (#14)

+++ BankApp/Entities/AccountType.cs
@@ -9,5 +9,5 @@ public class AccountType
     [Key] public short Id { get; set; }
     public string Code { get; set; } = string.Empty;
     public string Name { get; set; } = string.Empty;
-    public float InterestRate { get; set; }
+    public decimal InterestRate { get; set; }
 }
